no idea if ive done this right, moving functions and types to different directories
