+++ src/types/EventData.ts
@@ -1,0 +1,11 @@
+export type EventData = {
+    _id: string;
+    title: string;
+    day: string;
+    month: string;
+    date: string;
+    image: string;
+    locked: boolean;
+    disabled: boolean;
+  };
+
